pyproject.toml
[project]
name = "impact-scan"
version = "0.3.0"
description = "AI-powered security vulnerability scanner for codebases with intelligent fix suggestions"
authors = [
    {name = "Anirudh", email = "anirudh.ashrith2005@gmail.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
homepage = "https://github.com/Ani07-05/impact-scan"
repository = "https://github.com/Ani07-05/impact-scan"
documentation = "https://github.com/Ani07-05/impact-scan#readme"
keywords = [
    "security",
    "vulnerability",
    "scanner",
    "ai",
    "static-analysis",
    "dependency-audit",
    "sarif"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10", 
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.9,<4.0"
dependencies = [
    "pathspec>=0.12.1",
    "openai>=1.0.0",
    "anthropic>=0.28.0",
    "google-generativeai>=0.5.0",
    "google-genai>=1.0.0",
    "sarif-om>=1.0.4",
    "rich>=13.0.0",
    "textual>=0.41.0",
    "typer>=0.12.0",
    "bandit>=1.7.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "httpx>=0.25.0",
    "radon (>=6.0.1,<7.0.0)",
    "playwright>=1.40.0",
    "aiofiles>=23.2.0",
    "crawlee>=0.3.0",
    "pygments>=2.15.0",
    "pyyaml>=6.0.0",
    "markdown>=3.5.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "semgrep (>=1.134.0,<2.0.0)",
    "safety (>=3.6.1,<4.0.0)",
    "reportlab (>=4.4.3,<5.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "pytermgui (>=7.7.4,<8.0.0)",
    "groq (>=0.36.0,<0.37.0)",
    "packaging (>=25.0,<26.0)",
    "keyring (>=25.7.0,<26.0.0)",
]

[project.optional-dependencies]
local-llm = [
    "llama-cpp-python>=0.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
all = [
    "llama-cpp-python>=0.2.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]


[project.urls]
"Bug Tracker" = "https://github.com/Ani07-05/impact-scan/issues"
"Changelog" = "https://github.com/Ani07-05/impact-scan/releases"
"Discussions" = "https://github.com/Ani07-05/impact-scan/discussions"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/impact_scan"]

[tool.hatch.build.targets.wheel.force-include]
"src/impact_scan/templates" = "impact_scan/templates"
"src/impact_scan/rules" = "impact_scan/rules"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = "src"

[tool.pytest_anyio]
backend = "asyncio"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.1"
pip-audit = "^2.9.0"
ruff = "^0.14.6"
vulture = "^2.14"
pytest-textual-snapshot = "^1.1.0"
pytest-asyncio = {version = ">=0.21.0", markers = "python_version >= \"3.10\""}

[project.scripts]
impact-scan = "impact_scan.cli:app"
